requirements.txt
streamlit==1.35.0
requests==2.32.3
beautifulsoup4==4.12.3
pyttsx3==2.90
pyaudio==0.2.14
transformers==4.41.2
sentence-transformers==2.7.0
torch==2.3.0
chromadb==0.5.0
Pillow==10.3.0
tqdm==4.66.4
playwright==1.43.0
google-generativeai==0.5.4
